Refactor de la gestion des transactions et de l'authentification
- Ajout de AppTransactionService pour gérer toutes les opérations liées aux transactions :
    • Traitement des transferts entre utilisateurs
    • Récupération de l'historique des transactions (avec ou sans pagination)
    • Comptage des transactions pour la pagination

- Ajout de AuthenticationService pour centraliser la récupération de l'utilisateur authentifié.

- Modification de AppUserService :
    • Suppression de la logique liée aux transactions (déléguée à AppTransactionService)
    • Utilisation de AuthenticationService pour obtenir l'utilisateur courant
    • Maintien de la création, mise à jour et gestion des amis

- Modification de TransferController :
    • Délégation de toutes les opérations de transaction à AppTransactionService
    • Maintien de la récupération des amis via AppUserService
    • Conservation de la pagination, gestion des erreurs et binding du formulaire
